spa: Generalise property parsing
We can't have this be symmetric with the builder, because the number and
types of keys can't be known ahead of time.

The API still doesn't feel *great*, but we should be able to build on
top. Maybe a find() type API, plus some macros like we have in C.