spa: Implement a SPA plugin loader and log interface
This adds SPA plugin loading and refines the plugin interface so we can
actually run some code from a C plugin. There are a bunch of
indirections via Boxes here that we might be able to improve upon.

Unlike the C code, the loader lives in SPA for now, since that feels
more correct. Might need to move things around, especially since
registry bits will likely live in the PipeWire code rather than in SPA
code, but maybe we can achieve that separation without having to move
the loader code.

Also implement the Log interface and write a test so we can get this
going. The log interface probable needs some macros for actual usability
in code, though.