spa: support: Drop named interface setters
We move to a map of Box<dyn Interface>, and make sure we know how to
make/free native interface pointers from each. This takes us a bit
closer in the Rust support to the C version, and likely needs a little
more work for things to work like spa_support.